+++ src/pt2-clone/pt2_structs.h
@@ -102,6 +102,7 @@ typedef struct module_t
 
 	// for MOD2WAV progress bar
 	uint32_t rowsCounter, rowsInTotal;
+	float songDuration; // duration in seconds
 } module_t;
 
 typedef struct editor_t
@@ -144,15 +145,7 @@ typedef struct editor_t
 #endif
 } editor_t;
 
-typedef struct diskop_t
-{
-	volatile bool cached, isFilling, forceStopReading;
-	bool modPackFlg;
-	int8_t mode, smpSaveType;
-	int32_t numEntries, scrollOffset;
-} diskop_t;
 extern editor_t editor;
-extern diskop_t diskop;
 extern module_t *song; // pt2_main.c
 
 #ifdef __cplusplus